requirements.txt
discord.py==1.7.2
toml
git+https://github.com/Gorialis/jishaku
psutil